Changed from handlers to controllers, started to separate the project in layers, have to implement the postgres repository
