--- app/handlers/home.go
@@ -1,22 +1,0 @@
-package handlers
-
-import (
-	"encoding/json"
-	"net/http"
-)
-
-type HomeResponse struct {
-	Message string `json:"message"`
-	Status  bool   `json:"status"`
-}
-
-func HomeHandler(w http.ResponseWriter, r *http.Request) {
-	//W will send the response
-	w.Header().Add("Content-Type", "application/json")
-	//Http response
-	w.WriteHeader(http.StatusOK)
-	json.NewEncoder(w).Encode(HomeResponse{
-		Message: "Hello World",
-		Status:  true,
-	})
-}
